% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGSI.r
\name{AGSI}
\alias{AGSI}
\title{Accumulated growing season index model
as defined by Xin et al. 2015 (Rem. Sens. Env.)}
\usage{
AGSI(par, data)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{par:}{a vector of parameter values, this is functions specific}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Accumulated growing season index model
as defined by Xin et al. 2015 (Rem. Sens. Env.)
}
\examples{

\dontrun{
estimate = AGSI(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
