% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_plot.r
\name{arrow_plot}
\alias{arrow_plot}
\title{Generates an arrow plot, plotting differences
between two model runs visualized by red and
blue arrows indicating the direction of the change.}
\usage{
arrow_plot(data, models = NULL, lwd = 1.3, length = 0.03)
}
\arguments{
\item{data}{a structure list generated by model_comparison()}

\item{models}{models to compare, vector with two model names e.g.
c("TT","PTT") (default, first two models in the data file)}

\item{lwd}{line witdth of the arrows in the plot, change for readability
(default = 1.3)}

\item{length}{length of the arrow head, change for readability
(default = 0.03)}
}
\description{
This allows you to explore the effect of model structure
on the estimated phenology dates.
}
\examples{

\dontrun{
arrow_plot(data = model_comparison_data)
}
}
\keyword{comparison,}
\keyword{model}
\keyword{phenology,}
\keyword{plotting,}
\keyword{visualization}
