% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AT.r
\name{AT}
\alias{AT}
\title{Alternating model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
AT(par, data, plot = TRUE)
}
\arguments{
\item{data:}{a nested list of data with on location:
1. the date (doy or long format)
2. the temperature data
3. the photoperiod data (NA when not needed)
4. a vector or matrix with necessary constants (NA when not needed)
   - long term mean temperature
   - latitude
   - etc...}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Alternating model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = TT(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
