% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AT_Jan1_5C.r
\name{AT_Jan1_5C}
\alias{AT_Jan1_5C}
\title{Alternating model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with fixed January 1 start date and 5 deg C base temperature}
\usage{
AT_Jan1_5C(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Alternating model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with fixed January 1 start date and 5 deg C base temperature
}
\examples{

\dontrun{
estimate = AT_Jan1_5C(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
