% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{CDD}
\alias{CDD}
\title{Chilling Degree Day (CDD) adapted from
Jeong & Medvigny 2014 (Global Ecology & Biogeography)}
\usage{
CDD(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Chilling Degree Day (CDD) adapted from
Jeong & Medvigny 2014 (Global Ecology & Biogeography)
}
\examples{

\dontrun{
estimate <- CDD(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
