% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npn_phenophase.r
\name{check_npn_phenophases}
\alias{check_npn_phenophases}
\title{Checks if USA-NPN phenophase exists}
\usage{
check_npn_phenophases(phenophase = NULL, list = TRUE)
}
\arguments{
\item{phenophase}{An USA-NPN phenophase (character or number)}

\item{list}{List all species numbers and names as verbose output}
}
\value{
a validated list of species numbers, if not a warning is thrown
and any depended routines halted.
}
\description{
Checks if USA-NPN phenophase exists
}
\examples{

\dontrun{
# list all USA-NPN phenophases
check_npn_phenophase(list = TRUE)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
