% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pep725_species.r
\name{check_pep725_species}
\alias{check_pep725_species}
\title{Checks if PEP725 species name or number exists or can be generated}
\usage{
check_pep725_species(species = NULL, list = FALSE)
}
\arguments{
\item{species}{A species to download, either specified by its
species number or species name.}

\item{list}{List all species numbers and names as verbose output}
}
\value{
a validated list of species numbers, if not a warning is thrown
and any depended routines halted.
}
\description{
Checks if PEP725 species name or number exists or can be generated
}
\examples{

\dontrun{
check_pep725_species(species = 115)
}

}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
