% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{CU}
\alias{CU}
\title{CU chilling degree model
Chen et al. 2017 (Journal of Ecology)}
\usage{
CU(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
CU chilling degree model
Chen et al. 2017 (Journal of Ecology)
}
\examples{

\dontrun{
estimate = CU(data = data, par = par)
}
}
\keyword{forest}
\keyword{model,}
\keyword{phenology,}
\keyword{tropical}
