% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.r
\name{cvmae}
\alias{cvmae}
\title{Coefficient of Variation of the Mean Absolute Error}
\usage{
cvmae(par, data, model, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{nested data structure with validation data as returned
by format_phenocam() or format_pep725(), or your own dataset adhering
to the same data structure.}

\item{model}{the model name to be used in optimizing the model}

\item{...}{extra arguments to pass to the function}
}
\value{
the CVMAE comparing observed and estimated values
}
\description{
Coefficient of Variation of the Mean Absolute Error
}
\examples{

# The cost function returns the CVMAE between the
# true values and those generated by the model given a
# parameterset par.
\dontrun{
cost_value = cvmae(par, data, model="TTs")
}
}
\keyword{cost}
\keyword{function}
\keyword{model,}
\keyword{optimization,}
\keyword{phenology,}
