% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_berkeley_earth.r
\name{download_berkeley_earth}
\alias{download_berkeley_earth}
\title{Download Berkeley Earth Gridded mean daily temperature data}
\usage{
download_berkeley_earth(path = "~", year = 2011)
}
\arguments{
\item{path}{a path where to save the gridded data}

\item{year}{year to process (requires year - 1 to be present)}
}
\value{
nothing is returned to the R working environment, files are
downloaded and stored on disk
}
\description{
Download Berkeley Earth Gridded mean daily temperature data
}
\examples{

# donwload all gridded data for year 2014
\dontrun{
be_data = download_berkeley_earth(year = 2011)
}
}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
