% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cmip5.r
\name{download_cmip5}
\alias{download_cmip5}
\title{Download NASA Earth Exchange Global Daily Downscaled Projections (NEX-GDDP)
https://nex.nasa.gov/nex/}
\usage{
download_cmip5(path = "~", year = 2016, model = "miroc5",
  scenario = "rcp85", variable = c("tasmin", "tasmax", "pr"))
}
\arguments{
\item{path}{a path where to save the gridded data}

\item{year}{year to process (also requests year - 1)}
}
\value{
nothing is returned to the R working environment, files are
downloaded and stored on disk
}
\description{
Download NASA Earth Exchange Global Daily Downscaled Projections (NEX-GDDP)
https://nex.nasa.gov/nex/
}
\examples{

# donwload all gridded data for year 2014
\dontrun{
download_cmip5(year = 2011)
}
}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
