% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_npn.r
\name{download_npn}
\alias{download_npn}
\title{Automatically download USA-NPN data using the API interface}
\usage{
download_npn(species = 3, phenophase = 371, start = "2000-01-01",
  end = "2017-01-01", extent = NULL, internal = TRUE, path = ".")
}
\arguments{
\item{species}{A species to download, either specified by its
species number or species name. list species numbers and names with
check_npn_species(list = TRUE)}

\item{phenophase}{A phenophase number to filter for, a list of phenophases
is provided by check_npn_phenophases()}

\item{start}{start date in YYYY-MM-DD format}

\item{end}{end date in YYYY-MM-DD format}

\item{extent}{geographic coordinates constraining the output, defined
as bottom left, top right c(lon1, lat1, lon2, lat2) if null returns all
data (default = NULL)}

\item{internal}{completes download internally in a temporary directory}

\item{path}{the path where to save the downloaded data}
}
\value{
will return a data farme for the selected species,
phenophase, temporal and spatial extent or save the data to an RDS file
}
\description{
Automatically download USA-NPN data using the API interface
}
\examples{
\dontrun{
download_npn(species = 3, internal = FALSE)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
