% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pep725.r
\name{download_pep725}
\alias{download_pep725}
\title{Download PEP725 data}
\usage{
download_pep725(email = NULL, password = NULL, credentials = NULL,
  species = 115, path = "~", internal = TRUE)
}
\arguments{
\item{email}{email used in creating your pep725 download login}

\item{password}{password as created for the pep725 download login}

\item{credentials}{file with your email and password credentials, this is
the preferred method of storing your credentials over the explicit email and
password parameters as these might be forgotten in scripts and publicly
uploaded / shared. The credentials file is a file which lists your email and
password on two separate lines (in this order).}

\item{species}{A species to download, either specified by its
species number or species name. list species numbers and names with
check_pep725_species(list = TRUE)}

\item{path}{the path where to save the downloaded data}

\item{internal}{completes download internally in a temporary directory and
merges the data subsequently using merge_pep725(), returns a nested list
of tidy data. internal overrides the path command.}
}
\value{
will return csv files of PEP725 data for the selected species
}
\description{
Download PEP725 data
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
