% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.r
\name{DP}
\alias{DP}
\title{DormPhot model as defined in
Caffarra, Donnelly and Chuine 2011 (Clim. Res.)
parameter ranges are taken from Basler et al. 2016}
\usage{
DP(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
DormPhot model as defined in
Caffarra, Donnelly and Chuine 2011 (Clim. Res.)
parameter ranges are taken from Basler et al. 2016
}
\examples{

\dontrun{
estimate = DP(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
