% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{DU}
\alias{DU}
\title{DU drought model}
\usage{
DU(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Floral induction model by
Chen et al. 2017 (Journal of Ecology)
}
\examples{

\dontrun{
estimate <- DU(data = data, par = par)
}
}
\keyword{forest}
\keyword{model,}
\keyword{phenology,}
\keyword{tropical}
