% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_phenology.r
\name{estimate_phenology}
\alias{estimate_phenology}
\title{Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model. Wrapper around a do.call() call for convenience.}
\usage{
estimate_phenology(par, data, model)
}
\arguments{
\item{data:}{a nested list of data with on location:
1. the date (doy or long format)
2. the temperature data
3. the photoperiod data (NA when not needed)
4. a vector or matrix with necessary constants (NA when not needed)
   - long term mean temperature
   - latitude
   - etc...
5. validation data (optional when just running a model not optimizing)}

\item{par:}{a vector of parameter values, this is functions specific}

\item{model:}{the model name to be used in optimizing the model}
}
\description{
Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model. Wrapper around a do.call() call for convenience.
}
\examples{
estimate <- estimate.phenology(par,data,model)

# estimate will be an estimated timing of a phenophase
}
\keyword{model,}
\keyword{phenology,}
\keyword{post-processing}
