% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.DOY.chla.phenology.R
\name{extract.DOY.chla.phenology}
\alias{extract.DOY.chla.phenology}
\title{Extrait le jour de l'année du début du bloom printanier,
du jour de l'atteinte du pic de Chlorophylle-a
et sa valeur à partir de données extraites de
Google Earth Engine générées avec la fonction
ui.Chart.image.doySeriesByYear.}
\usage{
\method{extract}{DOY.chla.phenology}(filen, start = 32, end = 182,
  growth.rate.threshold = 0.01)
}
\arguments{
\item{filen}{est le nom du fichier CVS telechargé de GEE et produit par
ui.Chart.image.doySeriesByYear. Le fichier peut contenir plusieurs années.}

\item{start}{est le jour de l'année correspondant au debut de la période qui nous interesse
(valeur par défault est 32 (1 fevrier)).}

\item{end}{est le jour de l'année correspondant à la fin de la période qui nous interesse
(valeur par défault est 182 (1 juillet))}

\item{growth.rate.threshold}{est le taux de croissance à partir duquel on considère
le début du bloom. (valeur par défault est 0.01 mg Chl/m^3/jour)}
}
\value{
Retourne un tableau avec le jour de l'année marquant le début du bloom (bloom.onset),
le jour où le pic de Chla est atteint (DOY.peak.chla), la valeur du pic (peak.chla). et la durée de la
croissance du bloom (DC).
De plus, une figure en format *.png est crée dans le repertoire de travail pour
chaque année contenu dans le fichier CSV.
}
\description{
Extrait le jour de l'année du début du bloom printanier,
du jour de l'atteinte du pic de Chlorophylle-a
et sa valeur à partir de données extraites de
Google Earth Engine générées avec la fonction
ui.Chart.image.doySeriesByYear.
}
\author{
Simon Bélanger
}
