% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.DOY.NDSI.phenology.R
\name{extract.DOY.NDSI.phenology}
\alias{extract.DOY.NDSI.phenology}
\title{Extrait des événements phénologiques d'un cycle annuel de NDSI
à partir de données extraites de
Google Earth Engine générées avec la fonction
ui.Chart.image.doySeriesByYear.}
\usage{
\method{extract}{DOY.NDSI.phenology}(filen, Melt.Onset.Threshold = 0.75,
  Snow.Free.Threshold = 0.15, Fall.Snow.Thresold = 0.15)
}
\arguments{
\item{filen}{est le nom du fichier CVS telechargé de GEE et produit par
ui.Chart.image.doySeriesByYear. Le fichier peut contenir plusieurs années.}

\item{Melt.Onset.Threshold}{fraction de la valeur maximum de NDSI que l'on
considere en début de saison pour identifier le début de la fonte de la neige.
(valeur par défault est 0.75)}

\item{Snow.Free.Threshold}{fraction de la valeur maximum de NDSI que l'on
considere pour identifier la fonte complète de la neige.
(valeur par défault est 0.15)}

\item{Fall.Snow.Thresold}{fraction de la valeur maximum de NDSI que l'on
considere à la fin de l'été pour identifier l'apparition de la neige.
(valeur par défault est 0.15)}
}
\value{
Retourne un tableau avec le jour de l'année de chaque evenements.
De plus, une figure en format png est creee dans le répertoire de travail pour
chaque année contenue dans le fichier CSV.
}
\description{
La fonction trouve les dates (jour de l'année) de plusieurs événements phénologiques :
le jour du début de la fonte de la neige (Melt.Onset);
le jour de la fonte totale de la neige (Snow.Free);
le jour des premières neige d'automne (Fall.Snow);
la durée de la fonte (DF = Snow.Free - Melt.Onset);
la durée de la saison sans neige (DSSN= Fall.Snow - Snow.Free).
}
\author{
Simon Bélanger
}
