% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.DOY.peak.chla.R
\name{extract.DOY.peak.chla}
\alias{extract.DOY.peak.chla}
\title{Extrait le jour de l'annee du pic de Chlorophylle-a
et sa valeur à partir de donnees extraites de
Google Earth Engine generees avec la fonction
ui.Chart.image.doySeriesByYear.}
\usage{
extract.DOY.peak.chla(filen, start = 30, end = 180)
}
\arguments{
\item{filen}{est le nom du fichier CVS telecharger de GEE et produit par
ui.Chart.image.doySeriesByYear. Le fichier peut contenir plusieurs annees.}

\item{start}{est le jour de l'annee correspondant au debut de la periode qui nous interesse}

\item{end}{est le jour de l'annee correspondant a la fin de la periode qui nous interesse}
}
\value{
Retourne un tableau avec le jour de l'annee ou le pic de Chla
est atteint et la valeur du pic.
De plus, une figure en format png est cree dans le repertoire de travail pour
chaque annee contenu dans le fichier CSV.
}
\description{
Extrait le jour de l'annee du pic de Chlorophylle-a
et sa valeur à partir de donnees extraites de
Google Earth Engine generees avec la fonction
ui.Chart.image.doySeriesByYear.
}
\author{
Simon Bélanger
}
