% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_daymet (Koen’s MacBook Pro's conflicted
%   copy 2017-05-14).r, R/format_daymet.r
\name{format_daymet}
\alias{format_daymet}
\alias{format_daymet}
\title{Preprocessing of Daymet tiled data into a format which can be ingested
by the models in phenor}
\usage{
format_daymet(path = ".", year = 2014, tile = 11935, offset = 264,
  internal = TRUE)

format_daymet(path = ".", year = 2014, tile = 11935, offset = 264,
  internal = TRUE)
}
\arguments{
\item{path:}{a path to tiled data}

\item{year:}{year to process (requires year - 1 to be present)}

\item{tile:}{daymet tile number}

\item{offset:}{offset of the time series in DOY (default = 264, sept 21)}

\item{path:}{a path to tiled data}

\item{year:}{year to process (requires year - 1 to be present)}

\item{tile:}{daymet tile number}

\item{offset:}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
Preprocessing of Daymet tiled data into a format which can be ingested
by the models in phenor

Preprocessing of Daymet tiled data into a format which can be ingested
by the models in phenor
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
