% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_daymet_tiles.r
\name{format_daymet_tiles}
\alias{format_daymet_tiles}
\title{Preprocessing of Daymet tiled data into a format which can be ingested
by the models in phenor}
\usage{
format_daymet_tiles(path = "~", year = 2014, tile = 11935, offset = 264,
  internal = TRUE)
}
\arguments{
\item{path}{a path to tiled data}

\item{year}{year to process (requires year - 1 to be present)}

\item{tile}{daymet tile number}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{internal}{return results as an R variable or save as an RDS file}
}
\value{
Data file adhering to the phenor modelling input formatting. Data
serves as input for model spatial model runs. The only required data input
is the mean daily temperature, precipitation and VPD can be included but
are not mandatory. Depending on the models used, including only temperature
will keep file sizes down and processing time lower.
}
\description{
Preprocessing of Daymet tiled data into a format which can be ingested
by the models in phenor
}
\examples{

# run with default settings
# looks for daymet average temperature
# data in your home directory. These data
# are calculated using daymet_tmean() from
# the daymetr package
\dontrun{
daymet_data = format_daymet_tiles()
}
}
\keyword{climatology}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing,}
