% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_eobs.r
\name{format_eobs}
\alias{format_eobs}
\title{Preprocessing E-OBS data as hosted by:
http://www.ecad.eu/}
\usage{
format_eobs(path = "~", year = 2014, offset = 264, resolution = 0.25,
  internal = TRUE)
}
\arguments{
\item{path:}{a path of the gridded data netCDF files}

\item{year:}{year to process (requires year - 1 to be present / downloaded)}

\item{offset:}{offset of the time series in DOY (default = 264, sept 21)}

\item{resolution:}{0.25 or 0.50 degree data, only the normal rectangular data
will be processed not the polar variety.
calculation overhead}

\item{internal:}{TRUE / FALSE, write data structure to file or not (as .rds)}
}
\value{
Returns spatial model input data from long term E-OBS modelled data.
This data can be run by models specified in the phenor package.
}
\description{
Haylock, M.R., N. Hofstra, A.M.G. Klein Tank, E.J. Klok, P.D.
Jones, M. New. 2008: A European daily high-resolution gridded dataset
of surface temperature and precipitation.
J. Geophys. Res (Atmospheres), 113, D20119
}
\details{
Please register before downloading data, unzip data into uncompressed
netCDF files before processing.
}
\examples{

\dontrun{
# run with default settings netCDF E-OBS files will be
# looked for in your home directory
eobs_data = format_eobs()
}
}
\keyword{climate}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing,}
