% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gee.r
\name{format_gee}
\alias{format_gee}
\title{Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc. Ingests MODIS phenology dates as ingested
through the python based Google Earth Engine (GEE) subset tool. This is a
fallback method in the absence of a working version of R MODISTools.
(https://khufkens.github.io/gee_subset/)}
\usage{
format_gee(path = tempdir(), phenophase = "Increase", offset = 264)
}
\arguments{
\item{path}{a path to GEE provided MCD12Q2 phenology dates}

\item{phenophase}{Phenological phase, Increase, Maximum,
Decrease or Minimum (default = Increase)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc. Ingests MODIS phenology dates as ingested
through the python based Google Earth Engine (GEE) subset tool. This is a
fallback method in the absence of a working version of R MODISTools.
(https://khufkens.github.io/gee_subset/)
}
\examples{

\dontrun{
modis_data = format_gee()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
