% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_npn.r
\name{format_npn}
\alias{format_npn}
\title{Preprocessing of USA-NPN data into a format which can be ingested
by the optimization routines. Data is aggregated on a individual_id basis,
rather than on the user defined site_id.}
\usage{
format_npn(data = NULL, phenophase = 371, offset = 264, internal = TRUE)
}
\arguments{
\item{data}{an USA-NPN data frame as returned by download_npn()}

\item{phenophase}{a phenophase to include as validation statistic}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{internal}{return data as structured list to R workspace or write
to RDS file (default = TRUE)}
}
\description{
Preprocessing of USA-NPN data into a format which can be ingested
by the optimization routines. Data is aggregated on a individual_id basis,
rather than on the user defined site_id.
}
\examples{

\dontrun{
npn_data = format_npn()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
