% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_phenocam.r
\name{format_phenocam}
\alias{format_phenocam}
\title{Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc. the original nested list is flattened
for speed.}
\usage{
format_phenocam(path = "~", direction = "rising", gcc_value = "gcc_90",
  threshold = 50, offset = 264)
}
\arguments{
\item{path}{a path to 1 or 3-day PhenoCam time series
(no validation checks will be done, so mixed files will lead to
mixed results!)}

\item{direction}{rising = spring, falling = autumn}

\item{gcc_value}{gcc_90, gcc_mean, gcc_50 etc.}

\item{threshold}{10, 25, 50 = default (threshold)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc. the original nested list is flattened
for speed.
}
\examples{

# run with default settings
# looks for transition date files derived
# through phenocamr in your home directory
# change the path to match your setup
\dontrun{
phenocam_data = format_phenocam()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
