% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR.r
\name{GR}
\alias{GR}
\title{Thermal Time model as defined in
Garcia-Mozo et al. 2009 (Agr. For. Meteorlogy)}
\usage{
GR(par, data)
}
\arguments{
\item{data:}{a nested list of data}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Thermal Time model as defined in
Garcia-Mozo et al. 2009 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = GR(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
