% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{GRP}
\alias{GRP}
\title{Thermal Time grassland pollen model which includes a pulse response
precipitatoin trigger as defined in
Garcia-Mozo et al. 2009 (Agr. For. Meteorlogy)}
\usage{
GRP(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Thermal Time grassland pollen model which includes a pulse response
precipitatoin trigger as defined in
Garcia-Mozo et al. 2009 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate <- GRP(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
