% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{igbp_10}
\alias{igbp_10}
\title{igbp_10}
\format{A raster layer with land cover density (0-1) for an igbp class:
\describe{
  \item{raster}{land cover density for igbp class 10}
}}
\usage{
igbp_10
}
\description{
Raster mask of land cover density counts in 1x1 degree cells for the
conterminous US (CONUS). This data was generated by the
land_cover_density() function and it's default settings. The same dataset
can be generated with this function, but inclusion of the data in the
package saves some time. For areas outside the listed classes and CONUS
the function will have to be run again including a new source map and
destination grid.
}
\keyword{datasets}
