% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{LIN}
\alias{LIN}
\title{Linear model}
\usage{
LIN(par, data, spring = c(60, 151))
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{spring}{a vector defining spring as a couple of DOY values
default is March, April, May or DOY 60 - 151}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate <- LIN(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
