% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1.r
\name{M1}
\alias{M1}
\title{M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
M1(par, data, plot = FALSE)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = TT(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
