% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_Jan1_5C.R
\name{M1_Jan1_5C}
\alias{M1_Jan1_5C}
\title{M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with fixed January 1 start date and 5 deg C base temperature}
\usage{
M1_Jan1_5C(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with fixed January 1 start date and 5 deg C base temperature
}
\examples{

\dontrun{
estimate = TT(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
