% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1s.r
\name{M1s}
\alias{M1s}
\title{M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with a sigmoidal temperature response (Kramer 1994)}
\usage{
M1s(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with a sigmoidal temperature response (Kramer 1994)
}
\examples{

\dontrun{
estimate = TT(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
