% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1s_Jan1.r
\name{M1s_Jan1}
\alias{M1s_Jan1}
\title{M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with a sigmoidal temperature response (Kramer 1994)
and fixed Jan 1 start date}
\usage{
M1s_Jan1(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
with a sigmoidal temperature response (Kramer 1994)
and fixed Jan 1 start date
}
\examples{

\dontrun{
estimate = M1s_Jan1(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
