% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calibration.r
\name{model_calibration}
\alias{model_calibration}
\title{Model validation routine to faciliate model development
and quick validation of a (new) model.}
\usage{
model_calibration(model = "TT", data = phenocam_DB,
  control = list(max.call = 2000),
  par_ranges = sprintf("\%s/extdata/parameter_ranges.csv",
  path.package("phenor")), plot = TRUE, ...)
}
\arguments{
\item{model}{the model name to be used in optimizing the model}

\item{data}{dataset generated by the format_phenocam() or
format_modis() routines, or adhering to the general model optimization
input format.}

\item{control}{list of control parameters to be passed to the optimizer}

\item{par_ranges}{a vector of starting parameter values (function specific)
defaults to the parameter ranges as provided with the current models
and set forth by Basler (2016)}

\item{plot}{TRUE / FALSE, plot model fit}

\item{...}{additional parameters to be passed to the model}
}
\description{
Model validation routine to faciliate model development
and quick validation of a (new) model.
}
\examples{

\dontrun{
model_calibration(model,par_ranges = "parameter_ranges.csv")

# estimate will return the best estimated parameter set given the
# validation data
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{validation}
