% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.r
\name{model_comparison}
\alias{model_comparison}
\title{Model comparison routine to faciliate model development
and quick comparisons of the skill of various models.}
\usage{
model_comparison(random_seeds = c(1, 12, 40), models = c("LIN", "TT", "TTs",
  "PTT", "PTTs", "M1", "M1s", "AT", "SQ", "SQb", "SM1", "SM1b", "PA", "PAb",
  "PM1", "PM1b", "UN", "UM1", "SGSI", "AGSI"), data = phenocam_DB,
  method = "GenSA", control = list(max.call = 5000, temperature = 10000),
  par_ranges = sprintf("\%s/extdata/parameter_ranges.csv",
  path.package("phenor")), ncores = 1)
}
\arguments{
\item{random_seeds}{a vector with random seeds for cross validation}

\item{models}{list of models to compare}

\item{data}{which standard or custom dataset to use}

\item{method}{which optimization method to use, GenSA or rgenoud
(default = GenSA)}

\item{control}{additional optimization control parameters
(default = list(max.call = 5000, temperature = 10000))}

\item{par_ranges}{location of the parameter ranges of the models}

\item{ncores}{number of cores to use to calculate model comparisons,
system specific and defaults to 1 threat (default = 1)}
}
\description{
Model comparison routine to faciliate model development
and quick comparisons of the skill of various models.
}
\examples{

# estimate will return the best estimated parameter set given the
# validation data
\dontrun{
my_comparison <- model_comparison(random_seeds = c(38,1),
 models = c("TT","PTT"),
 dataset = "phenocam_DB",
 par_ranges = "parameter_ranges.csv")
}
}
\keyword{comparison}
\keyword{model,}
\keyword{phenology,}
\keyword{validation,}
