% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cv.r
\name{model_cv}
\alias{model_cv}
\title{Model valiadation routine to faciliate model cross-validation (k-fold or leave-one-out).}
\usage{
model_cv(k = 10, cv_seed = 1234, cv_set = NULL, ncores = 1,
  random_seeds = c(1, 12, 40), models = c("LIN", "TT", "TTs", "PTT", "PTTs",
  "M1", "M1s", "AT", "SQ", "SQb", "SM1", "SM1b", "PA", "PAb", "PM1", "PM1b",
  "UN", "UM1", "SGSI", "AGSI"), data = phenocam_DB, method = "GenSA",
  control = list(max.call = 5000, temperature = 10000),
  par_ranges = sprintf("\%s/extdata/parameter_ranges.csv",
  path.package("phenor")))
}
\arguments{
\item{k}{perform k-fold cross validation
(set k=0 to perform a leave-one-out cross-validation)}

\item{cv_seed}{a vector with random seeds for cross validation
(to shuffle the input data)}

\item{cv_set}{(optional) use custom sets cross validation
(requires a list of vectors the indices of the data used for validation
(hold-out), the remaining data will be used for calibration)}

\item{ncores}{number of cores to use for crossvalidation
(requires 'doParallel' and 'foreach'  library)}

\item{random_seeds}{a vector with random seeds for optimization replicates}

\item{models}{list of models to compare}

\item{data}{which standard or custom dataset to use}

\item{method}{which optimization method to use, GenSA or rgenoud
(default = GenSA)}

\item{control}{additional optimization control parameters
(default = list(max.call = 5000, temperature = 10000))}

\item{par_ranges}{location of the parameter ranges of the models}
}
\description{
Model valiadation routine to faciliate model cross-validation (k-fold or leave-one-out).
}
\examples{

\dontrun{
 model_CV(k=10,CV_seed = 123, models =  c("LIN","TT","PTT") , data= phenocam_DB[c(1:10)])

# will return the data of each validation and some summarizing statistics
}
}
\keyword{comparison}
\keyword{model,}
\keyword{phenology,}
\keyword{validation,}
