% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.r
\name{model_validation}
\alias{model_validation}
\title{Model validation routine to faciliate model development
and quick validation of a (new) model.}
\usage{
model_validation(model = "TT", dataset = "phenocam_DB",
  control = list(max.call = 2000),
  par_ranges = sprintf("\%s/extdata/parameter_ranges.csv",
  path.package("phenor")))
}
\arguments{
\item{par_ranges:}{a vector of starting parameter values (function specific)
defaults to the parameter ranges as provided with the current models
and set forth by Basler (2016)}

\item{model:}{the model name to be used in optimizing the model}
}
\description{
Model validation routine to faciliate model development
and quick validation of a (new) model.
}
\examples{

\dontrun{
model_validation(model,par_ranges = "parameter_ranges.csv")

# estimate will return the best estimated parameter set given the
# validation data
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{validation}
