% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.model.parameters.r
\name{optimize.model.parameters}
\alias{optimize.model.parameters}
\title{Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model.}
\usage{
optimize.model.parameters(par = par, data = data, cost = cost.function,
  model = "SEQ1.3", method = "GenSA", lower = -Inf, upper = Inf,
  maxit = 100)
}
\arguments{
\item{par:}{a vector of starting parameter values (function specific)}

\item{data:}{a nested list of data with on location:
1. the date (doy or long format)
2. the temperature data
3. the photoperiod data (NA when not needed)
4. a vector or matrix with necessary constants (NA when not needed)
   - long term mean temperature
   - latitude
   - etc... 
5. validation data (REQUIRED)}

\item{cost:}{the cost function to use in the optimization, it should return
a RMSE or other value which needs to be minimized}

\item{model:}{the model name to be used in optimizing the model}

\item{method:}{optimization method to use
- GenSA :  Generalized Simulated Annealing algorithm (default)
- genoud : GENetic Optimization Using Derivatives
- those available in optim()}

\item{lower:}{lower limit of parameter values (function specific)}

\item{upper:}{upper limit of parameter values (function specific)}

\item{maxit:}{maximum number of iterations to run (if NULL until convergence)}
}
\description{
Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model.
}
\examples{
estimate <- estimate.phenology(par,data,model)

# estimate will return the best estimated parameter set given the 
# validation data
}
\keyword{annealing,}
\keyword{genoud,}
\keyword{model,}
\keyword{optim}
\keyword{optimization,}
\keyword{phenology,}
\keyword{simulated}

