% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameters.r
\name{optimize_parameters}
\alias{optimize_parameters}
\title{Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model.}
\usage{
optimize_parameters(par = NULL, data = data, cost = rmse, model = "TT",
  method = "GenSA", lower = NULL, upper = NULL, control = NULL)
}
\arguments{
\item{par:}{a vector of starting parameter values (function specific)}

\item{data:}{a nested list of data}

\item{cost:}{the cost function to use in the optimization, it should return
a RMSE or other value which needs to be minimized}

\item{model:}{the model name to be used in optimizing the model}

\item{method:}{optimization method to use
- GenSA :  Generalized Simulated Annealing algorithm (default)
- genoud : GENetic Optimization Using Derivatives}

\item{lower:}{lower limit of parameter values (function specific)}

\item{upper:}{upper limit of parameter values (function specific)}

\item{maxit:}{maximum number of iterations to run (if NULL until convergence)}
}
\description{
Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model.
}
\examples{

\dontrun{
estimate <- estimate.phenology(par,data,model)

# estimate will return the best estimated parameter set given the
# validation data
}
}
\keyword{annealing,}
\keyword{genoud,}
\keyword{model,}
\keyword{optim}
\keyword{optimization,}
\keyword{phenology,}
\keyword{simulated}
