% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{outlier_detection}
\alias{outlier_detection}
\title{Outlier detection of transition dates}
\usage{
outlier_detection(df)
}
\arguments{
\item{df}{a nested list phenor formatted data file}
}
\value{
a nested list phenor formatted data file with outliers removed
}
\description{
Uses the 30-day rule (Schaber and Badeck, 2002) to remove outliers
from visual observations records. This function should be considered
for use with either PEP725 or USANPN data.
}
