% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA.r
\name{PA}
\alias{PA}
\title{Parallel model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
PA(par, data)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Parallel model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = PA(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
