% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAb.r
\name{PAb}
\alias{PAb}
\title{Parallel model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
using a bell shaped chilling response}
\usage{
PAb(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Parallel model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
using a bell shaped chilling response
}
\examples{

\dontrun{
estimate = PAb(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
