% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{phenocam_DB_autumn}
\alias{phenocam_DB_autumn}
\title{phenocam_DB_autumn}
\format{
A structured list with 6 variables:
\describe{
  \item{site}{PhenoCam site name}
  \item{location}{geographic location}
  \item{doy}{DOY (1 - 365 as a vector)}
  \item{transition_dates}{budburst dates, as per 25 percent of the Gcc amplitude }
  \item{Ti}{mean daily temperature, matrix 365 rows x # site years columns}
  \item{Li}{daylength, matrix 365 rows x # site years columns}
}
}
\usage{
phenocam_DB_autumn
}
\description{
Data structure containing leaf senescence dates for all deciduous broadleaf
(DB) sites included in the first PhenoCam data release. These data
are described in detail in Richardson et al. (2017)
}
\keyword{datasets}
