% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{phenocam_GR}
\alias{phenocam_GR}
\title{phenocam_GR}
\format{A structured list with 6 variables:
\describe{
  \item{site}{PhenoCam site name}
  \item{location}{geographic location}
  \item{doy}{DOY (1 - 365 as a vector)}
  \item{transition_dates}{budburst dates, as per 25 percent of the Gcc amplitude }
  \item{Ti}{mean daily temperature, matrix 365 rows x # site years columns}
  \item{Li}{daylength, matrix 365 rows x # site years columns}
}}
\usage{
phenocam_GR
}
\description{
Data structure containing budburst dates for all grassland (GR)
sites included in the first PhenoCam data release. These data
are described in detail in Richardson et al. (2017)
}
\keyword{datasets}
