% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pr_fit}
\alias{plot.pr_fit}
\title{Plotting method for fit model output}
\usage{
\method{plot}{pr_fit}(data)
}
\arguments{
\item{data}{input data generated using the pr_fit() function}
}
\value{
a plot with fit statistics
}
\description{
Plotting method for fit model output
}
\keyword{accuracy}
\keyword{model,}
\keyword{phenology,}
