% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_calc_land_cover.R
\name{pr_calc_land_cover}
\alias{pr_calc_land_cover}
\title{Calculates land cover density}
\usage{
pr_calc_land_cover(lc_raster, dest_raster, lc_classes = c(1, 4, 5, 10),
  path = tempdir(), internal = FALSE)
}
\arguments{
\item{lc_raster}{a MCD12Q1 map in epsg:4326.}

\item{dest_raster}{area to calculate the statistics for, with a coarser
resolution than the 500m MCD12Q1 data}

\item{lc_classes}{land cover classes to calculate the statistics for.
only IGBP classes (1 - 16) are supported. Takes a vector e.g. c(1, 4)}

\item{path}{path to output your generated data to if not outputting
it to the R console / workspace, default = "~"}

\item{internal}{TRUE / FALSE, if true no files are written to disk and
a raster stack is returned to the R command line}
}
\description{
Use MODIS MCD12Q1 land cover input map (src_raster) and a coarser
destination raster (E-OBS, CMIP5) to calculate land cover representation
of these coarser cells.
}
\details{
The algorithm counts the number of occurences in a given pixel of
the destination raster. Both maps should be in EPSG:4326.
}
\examples{

# will return a land cover density map for
# the evergreen needleleaf class (1) and
# the deciduous broadleaf classs (4)
\dontrun{
lc_dens = pr_calc_land_cover(lc_classes = c(1, 4),
                             lc_raster = "~/MCD12Q1_igbp_map.tif"
                             dest_raster = "~/1_degree_lat_lon_map.tif")
}
}
\keyword{comparison}
\keyword{model,}
\keyword{phenology,}
\keyword{validation,}
