% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_calc_temp_sens.R
\name{pr_calc_temp_sens}
\alias{pr_calc_temp_sens}
\title{Calculates the temperature sensitivity of a model (and parameters)}
\usage{
pr_calc_temp_sens(par, data, model, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{a nested list of data formatted using one of the format_*()
functions}

\item{model}{the model name to be used in optimizing the model}

\item{...}{extra arguments to pass to the function}
}
\value{
returns a temperature sensitivity as days per degree celsius
}
\description{
Calculates the temperature sensitivity of a model (and parameters)
}
\examples{

\dontrun{

# model calibration with default parameters
# phenocam_DB, "TT" to estimate a parameter set
pars <- pr_fit()

# estimate temperatue sensitivity for a particular
# combination of data, parameters and model
sensitivity <- pr_calc_temp_sens(par = pars$par,
 data = phenocam_DB, model = "TT")

}
}
\keyword{model,}
\keyword{phenology,}
\keyword{post-processing}
