% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_dl_be.R
\name{pr_dl_be}
\alias{pr_dl_be}
\title{Download Berkeley Earth Gridded mean daily temperature data}
\usage{
pr_dl_be(path = tempdir(), year = 2011)
}
\arguments{
\item{path}{a path where to save the gridded data}

\item{year}{year to process (requires year - 1 to be present)}
}
\value{
nothing is returned to the R working environment, files are
downloaded and stored on disk
}
\description{
Download Berkeley Earth Gridded mean daily temperature data
}
\examples{

# donwload all gridded data for year 2011
\dontrun{
pr_dl_be(year = 2011)
}
}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
