% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_dl_cmip5.r
\name{pr_dl_cmip5}
\alias{pr_dl_cmip5}
\title{Download NASA Earth Exchange Global Daily Downscaled Projections (NEX-GDDP)
https://nex.nasa.gov/nex/.}
\usage{
pr_dl_cmip5(
  path = tempdir(),
  year = 2016,
  model = "MIROC5",
  scenario = "rcp85",
  variable = c("tasmin", "tasmax", "pr")
)
}
\arguments{
\item{path}{a path where to save the gridded data}

\item{year}{year to process (also requests year - 1)}

\item{model}{CMIP5 model data to download (character vector)}

\item{scenario}{which RCP scenario to select for (default = "rcp85")}

\item{variable}{which climate variables to download (tasmin, tasmax, pr)
(default = c("tasmin","tasmax","pr"))}
}
\value{
nothing is returned to the R working environment, files are
downloaded and stored on disk
}
\description{
When downloading data for a particular year the antecedent year will
be downloaded as well as the format_cmip5() routine often requires a
matching dataset.
}
\details{
CMIP 5 models included are:
ACCESS1-0, CSIRO-MK3-6-0, MIROC-ESM, BCC-CSM1-1, GFDL-CM3, MIROC-ESM-CHEM,
BNU-ESM, GFDL-ESM2G, MIROC5, CanESM2, GFDL-ESM2M, MPI-ESM-LR, CCSM4, INM-CM4,
MPI-ESM-MR, CESM1-BGC, IPSL-CM5A-LR, MRI-CGCM3, CNRM-CM5,
IPSL-CM5A-MR, NorESM1-M
}
\examples{

# donwload all gridded data for year 2011 (and 2010)
# and format the data correctly.
\dontrun{
pr_dl_cmip5(year = 2011,
               path = tempdir(),
               model = "MIROC5",
               scenario = "rcp85")

cmip5_data <- pr_fm_cmip5(path = tempdir(),
                          offset = 264,
                          model = "MIROC5",
                          scenario = "rcp85",
                          year = 2011)
}
}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
