% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_dl_era5.R
\name{pr_dl_era5}
\alias{pr_dl_era5}
\title{Download Global ERA5 driver data}
\usage{
pr_dl_era5(
  path = tempdir(),
  file = "era5.nc",
  product = "era5",
  year = 2019,
  extent = c(-80, -70, 40, 50),
  user
)
}
\arguments{
\item{path}{a path where to save the gridded data}

\item{file}{filename of the file holding the final ERA5 netcdf data}

\item{product}{which ERA5 product to download,}

\item{year}{year of the data queried}

\item{extent}{vector with coordinates defining the region of interest defined
as ymax, xmin, ymin, xmax in lat/lon (default = c( 40, -80, 50, -70))}

\item{user}{Copernicus Data Store user ID (a number), on linux do not forget
to set options(keyring_backend='file') if you use a file based keyring.}
}
\value{
nothing is returned to the R working environment, files are
downloaded and stored on disk
}
\description{
Data is downloaded from the Copernicus Data Store, this routine therefore
requires a valid account with this service and your ID and key installed on
your system using the 'ecmwfr' package.
}
\details{
The routine provides access to all available model runs and is a user
friendly wrapper around a standard 'ecmwfr' call. Custom data formatting is
possible, but herefore we refer to the 'ecmwfr' package.
}
\examples{

\dontrun{
print("todo")
}
}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
