% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fit_parameters.R
\name{pr_fit_parameters}
\alias{pr_fit_parameters}
\title{Fit parameters of a phenology model to validation data}
\usage{
pr_fit_parameters(par = NULL, data, cost = rmse, model = "TT",
  method = "GenSA", lower, upper, control, ...)
}
\arguments{
\item{par}{a vector of starting parameter values (function specific)}

\item{data}{nested data structure with validation data as returned
by format_phenocam() or format_pep725(), or your own dataset adhering
to the same data structure.}

\item{cost}{the cost function to use in the optimization, it should return
a RMSE or other value which needs to be minimized}

\item{model}{the model name to be used in optimizing the model}

\item{method}{optimization method to use (default = GenSA)
- GenSA :  Generalized Simulated Annealing algorithm
- genoud : GENetic Optimization Using Derivatives
- BayesianTools: various bayesian based optimization tools}

\item{lower}{lower limit of parameter values (function specific)}

\item{upper}{upper limit of parameter values (function specific)}

\item{control}{optimization control parameters (default = NULL)}

\item{...}{extra arguments to pass to the function, mostly BayesianTools}
}
\description{
Calculates the estimated phenophases for a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model.
}
\examples{

\dontrun{
estimate <- pr_fit_parameters(par,data,model)

# estimate will return the best estimated parameter set given the
# validation data
}
}
\keyword{annealing,}
\keyword{genoud,}
\keyword{model,}
\keyword{optim}
\keyword{optimization,}
\keyword{phenology,}
\keyword{simulated}
