% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_flatten.R
\name{pr_flatten}
\alias{pr_flatten}
\title{Generates a flattened 'phenor' data file}
\usage{
pr_flatten(data)
}
\arguments{
\item{data}{structure list generated by pr_fm_*() functions}
}
\value{
returns a flat file format structure of the pr_fm_*() input
data, this to speed up processing
}
\description{
Uses files generated by pr_fm_*()
Flattening the file format allows for substantial speed increases
in optimization however limits readability. Using the split functionality
between the pr_fm_*() functions and this function allows for easier
subsetting of datasets.
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
