% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fm_cmip.R
\name{pr_fm_cmip}
\alias{pr_fm_cmip}
\title{Formatting CMIP6 data}
\usage{
pr_fm_cmip(
  path = tempdir(),
  year = 2016,
  offset = 264,
  model = "MIROC6",
  scenario = "ssp5_8_5",
  extent = c(40, -80, 50, -70),
  internal = TRUE
)
}
\arguments{
\item{path}{a path to the gridded data}

\item{year}{year to process (requires year - 1 to be present / downloaded)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{model}{which CMIP6 model to use}

\item{scenario}{"rcp85", "rcp45", "historical" here rcp covers 2006 - 2100
while historical data covers 1950 - 2005}

\item{extent}{vector with coordinates defining the region of interest defined
as xmin, xmax, ymin, ymax in lat/lon (default = c(-80, -70, 40, 50)), note
that this differs from the download routine!!}

\item{internal}{TRUE / FALSE, write data structure to file as RDS
(default = FALSE)}
}
\value{
data structure formatted for phenor model optimization and validation
}
\description{
Formats CMIP6 data as downloaded through the Copernicus CDS service,
using pr_dl_cmip().
}
\examples{
# run with default settings
# look for alternative models on the CMIP6
# downscaled data page
\dontrun{
cmip5_data <- pr_fm_cmip()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
