% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fm_gee.R
\name{pr_fm_gee}
\alias{pr_fm_gee}
\title{Preprocessing of GEE data}
\usage{
pr_fm_gee(path = tempdir(), phenophase = "Increase", offset = 264)
}
\arguments{
\item{path}{a path to GEE provided MCD12Q2 phenology dates}

\item{phenophase}{Phenological phase, Increase, Maximum,
Decrease or Minimum (default = Increase)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
Ingests MODIS phenology dates as ingested
through the python based Google Earth Engine (GEE) subset tool. This is a
fallback method in the absence of a working version of R MODISTools.
(https://khufkens.github.io/gee_subset/)
}
\examples{

\dontrun{
modis_data = pr_fm_gee()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
