% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fm_pep725.R
\name{pr_fm_pep725}
\alias{pr_fm_pep725}
\title{Preprocessing of PEP725 data into a format which can be ingested
by the optimization routines.}
\usage{
pr_fm_pep725(pep_path = tempdir(), eobs_path = tempdir(),
  bbch = "11", species = NULL, offset = 264, count = 60,
  resolution = 0.25, pep_data)
}
\arguments{
\item{pep_path}{path to the PEP725 data (species files only)}

\item{eobs_path}{path to regular grid E-OBS data.}

\item{bbch}{which phenophase (bbch) to use (default = 11)}

\item{species}{species to select from merged PEP725 file}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{count}{minimum number of acquisitions per location}

\item{resolution}{resolution of the E-OBS data (0.25 or 0.5, default = 0.25)}

\item{pep_data}{prefiltered subset of the merged PEP725 data obtained through
merge_PEP725(), else the complete data contained in pep_path will be used}
}
\value{
returns a nested list of site locations, their respective
phenological metrics and matching environmental data as extracted from
the E-OBS product (corrected for altitude using a lapse rate of 5C/km.)
}
\description{
Some pre-processing steps are required as downloading the PEP725.
So, for the species of interest download the separate zipped files.
Unzip all files and put the respective scientific data in one folder.
}
\details{
The routine requires E-OBS climate, which can be downloaded from
the E-OBS website and should be put unzipped in a single folder.
(http://www.ecad.eu/download/ensembles/ensembles.php).
}
\examples{

# run with default settings
# looks for transition date files derived
# through phenocamr in your home directory
# change the path to match your setup
\dontrun{
phenocam_data = format_pep725(pep_path = "~/pep_data/",
                              eobs_path = "~/eobs_data/")
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
