% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fm_subset.R
\name{pr_fm_subset}
\alias{pr_fm_subset}
\title{Subset formatted phenor data}
\usage{
pr_fm_subset(data, selection)
}
\arguments{
\item{data}{structured data of type/class 'phenor', generated by the phenor
format_...() functions}

\item{selection}{vector of columns to select (logical) of the same length
as the number of site years in the dataset}
}
\value{
Returns a subset of the original data, for cross validation
experiments.
}
\description{
Function to subset data formatted using the phenor package
}
\examples{

\dontrun{
# subset phenor formatted data
subset <- pr_fm_subset(data = phenocam_DB,
                        selection = 1:10)
}
}
\keyword{climate}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing,}
