% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_merge_pep725.R
\name{pr_merge_pep725}
\alias{pr_merge_pep725}
\title{Merge PEP725 data files}
\usage{
pr_merge_pep725(path = tempdir())
}
\arguments{
\item{path}{a path to the PEP725 data (either a directory containing tar.gz
files or a single tar.gz file)}
}
\value{
concatted data of all data in the path as a tidy data frame
including all normal parameters, and the species name and country code
as derived from the file name
}
\description{
Preprocessing of PEP725 data, merges separate files into tidy
data, with each observation a line, each column a different
parameter value.
}
\examples{

\dontrun{
tidy_pep_data <- pr_merge_pep725()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
