% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pr_parameters}
\alias{pr_parameters}
\title{Read parameter boundary values}
\usage{
pr_parameters(
  model,
  par_ranges = system.file("extdata", "parameter_ranges.csv", package = "phenor",
    mustWork = TRUE)
)
}
\arguments{
\item{model}{a phenology model name}

\item{par_ranges}{location of the model parameter boundary file}
}
\description{
Read parameter boundary values
}
\examples{

T_response = triangular_temperature_response(T = 0:45)
\dontrun{
plot(0:45, T_response, type = "l")
}
}
