% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_predict.R
\name{pr_predict}
\alias{pr_predict}
\title{Predicts phenophases}
\usage{
pr_predict(par, data, model = "TT", path = ".", reproject = TRUE, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{a nested list of data formatted using one of the format_*()
functions}

\item{model}{the model name to be used in optimizing the model}

\item{path}{the path to daymet tile data formated using the
format_daymet() function. All tiles in this directory will be
processed with the specified model and data mosaicked and reprojected
if requested}

\item{reproject}{TRUE / FALSE, reproject to lat-lon (default = TRUE)}

\item{...}{extra arguments to pass to the function}
}
\value{
returns point based or gridded estimates of a particular phenopase
}
\description{
Uses a given set of parameters
and a specified model (be sure to match parameter and requirements
with the model. Wrapper around a do.call() call for convenience.
}
\examples{

\dontrun{
# estimate will be an estimated timing of a phenophase
estimate <- pr_predict(par, data, model)

# if a path is specified and no other data all tiled
# data in this directory will be processed into a map
# if a dataset is provided at the data parameter side
# only this particular tile will be processed.

}
}
\keyword{model,}
\keyword{phenology,}
\keyword{post-processing}
