% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_modis.r
\name{process_modis}
\alias{process_modis}
\title{Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc.}
\usage{
process_modis(path = "~", direction = "Increase", offset = 264)
}
\arguments{
\item{path:}{a path to MODISTools MCD12Q2 phenology dates}

\item{direction:}{Increase (= default), Maximum, Decrease, Minimum}

\item{offset:}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc.
}
\examples{

\dontrun{
# run with default settings
modis_data = format_modis()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
