% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTT_Jan1.r
\name{PTT_Jan1}
\alias{PTT_Jan1}
\title{Photothermal Time model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
PTT_Jan1(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Photothermal Time model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = PTT_Jan1(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
