% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.r
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error cost function for model optimization.}
\usage{
rmse(par, data, model, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{nested data structure with validation data as returned
by format_phenocam() or format_pep725(), or your own dataset adhering
to the same data structure.}

\item{model}{the model name to be used in optimizing the model}

\item{...}{extra arguments to pass to the function}
}
\value{
the RMSE comparing observed and estimated values
}
\description{
Root Mean Squared Error cost function for model optimization.
}
\examples{

# The cost function returns the rmse between the
# true values and those generated by the model given a
# parameterset par.
\dontrun{
cost_value = rmse(par, data, model="TTs")
}
}
\keyword{cost}
\keyword{function}
\keyword{model,}
\keyword{optimization,}
\keyword{phenology,}
