% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rotate_cmip}
\alias{rotate_cmip}
\title{Rotate CMIP NetCDF data cubes}
\usage{
rotate_cmip(r = NULL, extent = c(-74, -65, 40, 48))
}
\arguments{
\item{r}{CMIP raster layer, stack or brick}

\item{extent}{vector with coordinates defining the region of interest defined
as xmin, xmax, ymin, ymax in lat/lon (default = c(-74,-65, 40, 48))}
}
\value{
Returns raster data in the same format (layer, stack, brick) as
the input, but rotated from a climatological longitude (0 - 360) format to
a normal longitude format (-180 to 180).
}
\description{
Rotate NetCDF files faster than the default
raster rotate() command, as data is cropped
before any translations (reducing memory load)
}
