% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{SM1b}
\alias{SM1b}
\title{Sequential model (M1 variant) as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
using a bell shaped chilling response}
\usage{
SM1b(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Sequential model (M1 variant) as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
using a bell shaped chilling response
}
\examples{

\dontrun{
estimate <- SM1b(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
