% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQ.R
\name{SQ}
\alias{SQ}
\title{Sequential model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
SQ(par, data)
}
\arguments{
\item{data:}{a nested list of data}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Sequential model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = SQ(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
