% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_phenor.r
\name{subset_phenor}
\alias{subset_phenor}
\title{Subset formatted phenor data}
\usage{
subset_phenor(data, selection)
}
\arguments{
\item{data}{structured data of type/class 'phenor', generated by the phenor
format_...() functions}

\item{selection}{vector of columns to select (logical) of the same length
as the number of site years in the dataset}
}
\value{
Returns a subset of the original data, for cross validation
experiments.
}
\description{
Function to subset data formatted using the phenor package
}
\examples{

\dontrun{
# subset phenor formatted data
subset <- subset_phenor()
}
}
\keyword{climate}
\keyword{data}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing,}
