% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.models.r
\name{SW1.1}
\alias{SW1.1}
\title{Calculates estimated phenophases for a set of common phenology
models as listed in Melaas et al. 2013. In particular it uses a
photoperiod constrain to account for changes in latitude instead
of explicitely changing parameter p0 (see paper).}
\usage{
SW1.1(par, data)
}
\arguments{
\item{data:}{a nested list of data with on location:
1. the date (doy or long format)
2. the temperature data
3. the photoperiod data (NA when not needed)
4. a vector or matrix with necessary constants (NA when not needed)
   - long term mean temperature
   - latitude
   - etc...}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Calculates estimated phenophases for a set of common phenology
models as listed in Melaas et al. 2013. In particular it uses a
photoperiod constrain to account for changes in latitude instead
of explicitely changing parameter p0 (see paper).
}
\examples{
estimate <- SEQ1.3(data,par)

# estimate will be an estimated timing of a phenophase
# the format depends on the date as forwarded into the
# model

}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}

