% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_sensitivity.R
\name{temperature_sensitivity}
\alias{temperature_sensitivity}
\title{Calculates the temperature sensitivity of a model (and parameters)}
\usage{
temperature_sensitivity(par, data, model, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{a nested list of data formatted using one of the format_*()
functions}

\item{model}{the model name to be used in optimizing the model}

\item{...}{extra arguments to pass to the function}
}
\value{
returns a temperature sensitivity as days per degree celsius
}
\description{
Calculates the temperature sensitivity of a model (and parameters)
}
\examples{

\dontrun{

# model calibration with default parameters
# phenocam_DB, "TT" to estimate a parameter set
pars <- model_calibration()

# estimate temperatue sensitivity for a particular
# combination of data, parameters and model
sensitivity <- temperature_sensitivity(par = pars$par,
 data = phenocam_DB, model = "TT")

}
}
\keyword{model,}
\keyword{phenology,}
\keyword{post-processing}
