% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{triangular_temperature_response}
\alias{triangular_temperature_response}
\title{Triangular temperature response function}
\usage{
triangular_temperature_response(T = -10:45, T_opt = 10, T_min = 1, T_max = 15)
}
\arguments{
\item{T}{a vector or matrix of temperatures}

\item{T_opt}{optimal temperature}

\item{T_min}{minimum viable temperature}

\item{T_max}{maximum viable temperature}
}
\description{
As defined in: Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

T_response = triangular_temperature_response(T = 0:45)
\dontrun{
plot(0:45, T_response, type = "l")
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{response}
\keyword{temperature}
