% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT.r
\name{TT}
\alias{TT}
\title{Thermal Time model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
TT(par, data)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{par:}{a vector of parameter values, this is functions specific}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Thermal Time model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = TT(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
