% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UM1.r
\name{UM1}
\alias{UM1}
\title{Unified M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
UM1(par, data)
}
\arguments{
\item{data:}{a nested list of data}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Unified M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = UM1(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
