% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{UN}
\alias{UN}
\title{Unified model}
\usage{
UN(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
as defined in Chuine 2000
}
\examples{

\dontrun{
estimate <- UN(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
