% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrss.R
\name{wrss}
\alias{wrss}
\title{Calculates the weighted Residual Sum of Squares (wRSS) as a cost function for
model optimization. Different cost functions can be specified
if they adhere to the same basic structure in their parameters.
Currently only wRSS is provided as a cost function,}
\usage{
wrss(par, data, model, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{nested data structure with validation data as returned
by format_phenocam() or format_pep725(), or your own dataset adhering
to the same data structure.}

\item{model}{the model name to be used in optimizing the model
Most models are listed in Melaas et al. 2013, additional ones
will be added over time, preliminary list:
1. SEQ1.3}

\item{...}{extra arguments to pass to the function}
}
\value{
the wRSS comparing observed and estimated values
}
\description{
Calculates the weighted Residual Sum of Squares (wRSS) as a cost function for
model optimization. Different cost functions can be specified
if they adhere to the same basic structure in their parameters.
Currently only wRSS is provided as a cost function,
}
\examples{

# The cost function returns the wRSS between the
# true values and those generated by the model given a
# parameterset par.
\dontrun{
cost_value = wrss(par, data, model="TTs")
}
}
\keyword{cost}
\keyword{function}
\keyword{model,}
\keyword{optimization,}
\keyword{phenology,}
