% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{AGSI}
\alias{AGSI}
\title{Accumulated growing season index model (1)
as defined by Xin et al. 2015 (Rem. Sens. Env.)}
\usage{
AGSI(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
The starting point of accumulation is not clearly indicated
in the publication so we assume December 21.
}
\examples{

\dontrun{
estimate <- AGSI(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
