% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_plot.r
\name{arrow_plot}
\alias{arrow_plot}
\title{Generates an arrow plot, plotting differences
between two model runs visualized by red and
blue arrows indicating the direction of the change.}
\usage{
arrow_plot(data = NULL, models = NULL)
}
\arguments{
\item{data}{a structure list generated by model_comparison()}

\item{models}{models to compare, vector with two model names e.g.
c("TT","PTT")}
}
\description{
This allows you to explore the effect of model structure
on the estimated phenology dates.
}
\examples{

\dontrun{
arrow_plot(data = model_comparison_data)
}
}
\keyword{comparison,}
\keyword{model}
\keyword{phenology,}
\keyword{plotting,}
\keyword{visualization}
