% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AT.r
\name{AT}
\alias{AT}
\title{Alternating model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
AT(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Alternating model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = AT(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
