% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{daylength}
\alias{daylength}
\title{Calculates day length (in hours) and the solar elevation above the
ecliptic plane based upon latitude and a day of year, and year values.
according to H.Glarner (http://herbert.gandraxa.com/length_of_day.xml)}
\usage{
daylength(doy, latitude)
}
\arguments{
\item{doy}{a vector with doy values 1 - 365(6)}

\item{latitude}{a given latitude}
}
\value{
a daylength vector
}
\description{
Due to heterogeneous date formats with years normalized to 365 days
I do not apply a leap year correction.
}
\examples{

\dontrun{
# calcualte the hours of sunlight and solar elevation on day of year 1
length_of_day = daylength(1, 51, 2000)
print(length_of_day)
}
}
\keyword{ephemerids}
\keyword{solar,}
