% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pep725.r
\name{download_pep725}
\alias{download_pep725}
\title{Automatically download PEP725 data}
\usage{
download_pep725(credentials, species = 115, path = tempdir(),
  internal = FALSE)
}
\arguments{
\item{credentials}{Text file with your PEP725 email and password credentials.
The credentials file is a file which lists your email and
password on two separate lines (in this order).}

\item{species}{A species to download, either specified by its
species number or species name. list species numbers and names with
check_pep725_species(list = TRUE)}

\item{path}{the path where to save the downloaded data}

\item{internal}{completes download internally in a temporary directory and
merges the data subsequently using merge_pep725(), returns a nested list
of tidy data. internal overrides the path command.}
}
\value{
will return csv files of PEP725 data for the selected species
}
\description{
Uses the httr and rvest packages to scrape the PEP725 website for
phenology data. Login credentials are required so this does not
constitute a breach of the license agreement / data policy. Please
register here for access to the data:
http://www.pep725.eu/data_download/registration.php
and carefully read the data use policy:
http://www.pep725.eu/downloads/PEP_725_Data_Policy_201012.pdf
}
\examples{
\dontrun{
download_pep725(credentials = "~/pep725_login.txt",
                species = 115)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
