% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.DOY.peak.SST.warming.onset.R
\name{extract.DOY.peak.SST.warming.onset}
\alias{extract.DOY.peak.SST.warming.onset}
\title{Extrait le jour de l'année du début du réchauffement
des eaux de surface, du pic de SST
et sa valeur à partir de donnees extraites de
Google Earth Engine générées avec la fonction
ui.Chart.image.doySeriesByYear.}
\usage{
extract.DOY.peak.SST.warming.onset(filen)
}
\arguments{
\item{filen}{est le nom du fichier CVS telechargé de GEE et produit par
ui.Chart.image.doySeriesByYear. Le fichier peut contenir plusieurs annees.}
}
\value{
Retourne un tableau avec le jour de l'année ou le pic de Chla
est atteint et la valeur du pic.
De plus, une figure en format png est crée dans le répertoire de travail pour
chaque annee contenu dans le fichier CSV.
}
\description{
Extrait le jour de l'année du début du réchauffement
des eaux de surface, du pic de SST
et sa valeur à partir de donnees extraites de
Google Earth Engine générées avec la fonction
ui.Chart.image.doySeriesByYear.
}
\author{
Simon Bélanger
}
