% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_format.r
\name{flat_format}
\alias{flat_format}
\title{Flatten the format as generated by format_phenocam()
and format_modis(). Flattening the file format allows
for substantial speed increases in optimization however
limits readability. Using the split functionality between
the format_*() functions and this function allows for easy
subsetting of datasets.}
\usage{
flat_format(data = NULL)
}
\arguments{
\item{data:}{structure list generated by format_*() functions}
}
\description{
Flatten the format as generated by format_phenocam()
and format_modis(). Flattening the file format allows
for substantial speed increases in optimization however
limits readability. Using the split functionality between
the format_*() functions and this function allows for easy
subsetting of datasets.
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
