% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cmip5.r
\name{format_cmip5}
\alias{format_cmip5}
\title{Preprocessing of NASA Earth Exchange Global Daily Downscaled Projections
(NEX-GDDP).}
\usage{
format_cmip5(path = tempdir(), year = 2016, offset = 264,
  model = "IPSL-CM5A-MR", scenario = "rcp85", extent = c(-128, -65,
  24, 50), internal = TRUE)
}
\arguments{
\item{path}{a path to the gridded data (original max / min temp. / precip files)}

\item{year}{year to process (requires year - 1 to be present / downloaded)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{model}{which CMIP5 model to use}

\item{scenario}{"rcp85", "rcp45", "historical" here rcp covers 2006 - 2100
while historical data covers 1950 - 2005}

\item{extent}{vector with coordinates defining the region of interest defined
as xmin, xmax, ymin, ymax in lat/lon (default = c(-74,-65, 40, 48))}

\item{internal}{TRUE / FALSE, write data structure to file as RDS
(default = FALSE)}
}
\value{
data structure formatted for phenor model optimization and validation
}
\description{
Preprocessing of NASA Earth Exchange Global Daily Downscaled Projections
(NEX-GDDP).
}
\examples{
# run with default settings
# look for alternative models on the CMIP5
# downscaled data page
\dontrun{
cmip5_data = format_cmip5()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
