% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cmip5.r
\name{format_cmip5}
\alias{format_cmip5}
\title{Preprocessing CMIP5 model runs as hosted by:
http://gdo-dcp.ucllnl.org/downscaled_cmip_projections/dcpInterface.html}
\usage{
format_cmip5(path = "~", year = 2016, offset = 264,
  model = "IPSL-CM5A-MR", resolution = "1x1", scenario = "rcp85",
  internal = TRUE)
}
\arguments{
\item{path:}{a path to the gridded data (original max / min temp. files)}

\item{year:}{year to process (requires year - 1 to be present / downloaded)}

\item{offset:}{offset of the time series in DOY (default = 264, sept 21)}

\item{model:}{which CMIP5 model to use}

\item{resolution:}{"1x1" or "16th", "16th" will result in a significant
calculation overhead}

\item{scenario:}{"rcp85", "rcp45", "historical" here rcp covers 2006 - 2100
while historical data covers 1950 - 2005}

\item{internal:}{TRUE / FALSE, write data structure to file or not (as .rds)}
}
\description{
Although one can access the 1/16th degree data, by default the 1x1 degree
data should be preferred (due to preprocessing load).
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
