% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_csv.r
\name{format_csv}
\alias{format_csv}
\title{Pre-processing of CSV based data}
\usage{
format_csv(file = "~/Desktop/your_phenology.csv", phenophase,
  offset = 264, internal = TRUE, path = tempdir())
}
\arguments{
\item{file}{an CSV file with phenology observation dates}

\item{phenophase}{a phenophase to include as validation statistic}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{internal}{return data as structured list to R workspace or write
to RDS file (default = TRUE)}

\item{path}{output directory for converted data}
}
\description{
Combines CSV data into a format which can be ingested
by the optimization routines. The CSV requires columns named:
ID, lat, lon, phenophase, year, doy
}
\details{
subsets of the phenophases can be made, or are ignored if
not specified.
}
\examples{

\dontrun{
npn_data = format_csv()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
