% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_modis.r
\name{format_modis}
\alias{format_modis}
\title{Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc.}
\usage{
format_modis(path = "~", phenophase = "Increase", cycle = 1,
  offset = 264)
}
\arguments{
\item{path}{a path to MODISTools MCD12Q2 phenology dates}

\item{phenophase}{Phenological phase, Increase, Maximum,
Decrease or Minimum (default = Increase)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc.
}
\examples{

\dontrun{
modis_data = format_modis()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
