% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pep725.r
\name{format_pep725}
\alias{format_pep725}
\title{Preprocessing of PEP725 data into a format which can be ingested
by the optimization routines, the original nested list is flattened
for speed.}
\usage{
format_pep725(pep_path = "~/Downloads/PEP725_DE_111_000/",
  eobs_path = "~/tmp/eobs/", bbch = "11", species = NULL, offset = 264,
  count = 60, resolution = 0.25)
}
\arguments{
\item{path:}{a path to the PEP725 data (species files only)}

\item{bbch:}{which bbch}

\item{species:}{species to select from merged PEP725 file}

\item{offset:}{offset of the time series in DOY (default = 264, sept 21)}

\item{count:}{minimum number of acquisitions per location}
}
\value{
returns a nested list of site locations, their respective
phenological metrics and matching environmental data as extracted from
the E-OBS product (corrected for altitude using a lapse rate of xyz)
}
\description{
Some pre-processing steps are required as downloading the PEP725 data is a
mess (this database needs an API). So, for the species of interest
download the separate zipped files. Unzip all files and put the respective
scientific data in one folder.
}
\examples{

\dontrun{
# run with default settings
# looks for transition date files derived
# through phenocamr in your home directory
# change the path to match your setup
phenocam_data = format_pep725()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
