% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_phenocam.r
\name{format_phenocam}
\alias{format_phenocam}
\title{Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc. the original nested list is flattened
for speed.}
\usage{
format_phenocam(path = "~", direction = "rising", gcc_value = "gcc_90",
  threshold = 50, offset = 264, internal = TRUE)
}
\arguments{
\item{path}{a path to 1 or 3-day PhenoCam transition date estimates
(no validation checks will be done, so mixed files will lead to
mixed or duplicated results!)}

\item{direction}{The phenophase considered, "rising" for spring and
"falling" for autumn. When using fall the offset should be 365.
(default = rising)}

\item{gcc_value}{The gcc time series from which the phenophase estimates
were derived either (default = gcc_90)}

\item{threshold}{Threshold (% of seasonal amplitude) used in estimating
the phenophases (10, 25, 50) (default = 50)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{internal}{return data as structured list to R workspace or write
to RDS file (default = TRUE)}
}
\description{
Preprocessing of all PhenoCam data into a format which can be ingested
by the optimization routines etc. the original nested list is flattened
for speed.
}
\examples{

# run with default settings
# looks for transition date files derived
# through phenocamr in your home directory
# change the path to match your setup
\dontrun{
phenocam_data = format_phenocam()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
