% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR.r
\name{GR}
\alias{GR}
\title{Thermal Time grassland model which includes a pulse response
precipitatoin trigger as defined in
Garcia-Mozo et al. 2009 (Agr. For. Meteorlogy)}
\usage{
GR(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Thermal Time grassland model which includes a pulse response
precipitatoin trigger as defined in
Garcia-Mozo et al. 2009 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = GR(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
