% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.r
\name{likelihood}
\alias{likelihood}
\title{A likelihood cost function for model optimization.}
\usage{
likelihood(par, data, model, ...)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{nested data structure with validation data as returned
by format_phenocam() or format_pep725(), or your own dataset adhering
to the same data structure.}

\item{model}{the model name to be used in optimizing the model
Most models are listed in Melaas et al. 2013, additional ones
will be added over time, preliminary list:
1. SEQ1.3}

\item{...}{extra arguments to pass to the function}
}
\value{
the RMSE comparing observed and estimated values
}
\description{
A likelihood cost function for model optimization.
}
\examples{

# The cost function returns the rmse between the
# true values and those generated by the model given a
# parameterset par.
\dontrun{
cost_value = likelihood(par, data, model="TTs")
}

}
\keyword{cost}
\keyword{function}
\keyword{model,}
\keyword{optimization,}
\keyword{phenology,}
