% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{M1}
\alias{M1}
\title{M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
M1(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = M1(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
