% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pep725.r
\name{merge_pep725}
\alias{merge_pep725}
\title{Preprocessing of PEP725 data, merges separate files into tidy
data, with each observation a line, each column a different
parameter value.}
\usage{
merge_pep725(path = "~")
}
\arguments{
\item{path}{a path to the PEP725 data (species files only)}
}
\value{
concatted data of all data in the path as a tidy data frame
including all normal parameters, and the species name and country code
as derived from the file name
}
\description{
Preprocessing of PEP725 data, merges separate files into tidy
data, with each observation a line, each column a different
parameter value.
}
\examples{

\dontrun{
tidy_pep_data = merge_pep725()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
