% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.r
\name{model_comparison}
\alias{model_comparison}
\title{Model comparison routine to faciliate model development
and quick comparisons of the skill of various models.}
\usage{
model_comparison(random_seeds = c(1, 12, 40), models = c("LIN", "TT", "TTs",
  "PTT", "PTTs", "M1", "M1s", "AT", "SQ", "SQb", "SM1", "SM1b", "PA", "PAb",
  "PM1", "PM1b"), dataset = "phenocam_DB", method = "GenSA",
  control = list(max.call = 5000),
  par_ranges = sprintf("\%s/extdata/parameter_ranges.csv",
  path.package("phenor")))
}
\arguments{
\item{random_seeds:}{a vector with random seeds for cross validation}

\item{models:}{list of models to compare}

\item{dataset:}{which standard dataset to use}

\item{par_ranges:}{location of the parameter ranges of the models}
}
\description{
Model comparison routine to faciliate model development
and quick comparisons of the skill of various models.
}
\examples{

\dontrun{
my_comparison <- model_comparison(random_seeds = c(38,1),
 models = c("TT","PTT"),
 dataset = "phenocam_DB",
 par_ranges = "parameter_ranges.csv")

# estimate will return the best estimated parameter set given the
# validation data
}
}
\keyword{comparison}
\keyword{model,}
\keyword{phenology,}
\keyword{validation,}
