% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.r
\name{null}
\alias{null}
\title{Null model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
returns the mean across all validation dates}
\usage{
null(data)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\description{
Null model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
returns the mean across all validation dates
}
\examples{

\dontrun{
estimate = NLL(data = data)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
