% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_models.R
\name{PA}
\alias{PA}
\title{Parallel model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
PA(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
Parallel model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = PA(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
