% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_comparison.r
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Model comparison  plotting routine to faciliate model development
and quick comparisons of the skill of various models.}
\usage{
plot_model_comparison(data = NULL, ylab = "RMSE (days)", names = TRUE,
  ...)
}
\arguments{
\item{data}{list returned by model_comparison()}

\item{ylab}{cost function value to annotate y-axis (default = "RMSE (days)")}

\item{names}{include model names, TRUE / FALSE (default = TRUE)}

\item{...}{extra arguments to pass to the function}
}
\description{
Model comparison  plotting routine to faciliate model development
and quick comparisons of the skill of various models.
}
\examples{

\dontrun{
plot_model_comparison()
}
}
\keyword{comparison,}
\keyword{data,}
\keyword{model,}
\keyword{phenology,}
\keyword{plotting}
