% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM1.r
\name{PM1}
\alias{PM1}
\title{Parallel M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
PM1(par, data, plot = TRUE)
}
\arguments{
\item{data:}{a nested list of data}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Parallel M1 model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = PM1(data = data, par = par)
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
