% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fm_era5.R
\name{pr_fm_era5}
\alias{pr_fm_era5}
\title{Formatting ERA5 data}
\usage{
pr_fm_era5(
  path = tempdir(),
  year = 2016,
  offset = 264,
  extent = c(41, -79, 50, -69),
  internal = TRUE
)
}
\arguments{
\item{path}{a path to the gridded data}

\item{year}{year to process (requires year - 1 to be present / downloaded)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}

\item{extent}{vector with coordinates defining the region of interest defined
as xmin, xmax, ymin, ymax in lat/lon (default = c(-80, -70, 40, 50)), note
that this differs from the download routine!!}

\item{internal}{TRUE / FALSE, write data structure to file as RDS
(default = FALSE)}
}
\value{
data structure formatted for phenor model optimization and validation
}
\description{
Formats ERA5 data as downloaded through the Copernicus CDS service,
using pr_dl_era5().
}
\examples{
# run with default settings
# look for alternative models on the CMIP6
# downscaled data page
\dontrun{
cmip5_data <- pr_fm_cmip()
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{preprocessing}
