% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTT.r
\name{PTT}
\alias{PTT}
\title{Photothermal Time model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
PTT(par, data, plot = FALSE)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Photothermal Time model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = PTT(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
