% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.r
\name{rmse}
\alias{rmse}
\title{Calculates day length (in hours) and the solar elevation above the
ecliptic plane based upon latitude and a day of year value.}
\usage{
rmse(par, data, model, ...)
}
\arguments{
\item{par:}{a vector of parameter values, this is functions specific}

\item{data:}{a nested list of data with one location:
1. the date (doy or long format)
2. the temperature data
3. the photoperiod data (NA when not needed)
4. a vector or matrix with necessary constants (NA when not needed)
   - long term mean temperature
   - latitude
   - etc...
5. validation data (necessary!!)}

\item{model:}{the model name to be used in optimizing the model
Most models are listed in Melaas et al. 2013, additional ones
will be added over time, preliminary list:
1. SEQ1.3}
}
\description{
Different cost functions can be specified if they adhere to the
same basic structure in their parameters. Currently only RMSE is
provided as a cost function, defined as rmse().
}
\examples{

#' \dontrun{
cost_value = rmse(par,data,model="TTs")

# The cost function returns the rmse between the
# true values and those generated by the model given a
# parameterset par.
}
}
\keyword{model,}
\keyword{phenology,}
\keyword{sequential}
