% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGSI.r
\name{SGSI}
\alias{SGSI}
\title{Standard growing season index model
as defined by Xin et al. 2015 (Rem. Sens. Env.)}
\usage{
SGSI(par, data)
}
\arguments{
\item{par}{a vector of parameter values, this is functions specific}

\item{data}{input data (see reference for detailed description),
data should be formatted using flat_format()}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
No clear accumulation start dat was set in the above mentioned
manuscript, as such we assume a start date of 21th of Dec, or 1th of Jan.
depending on the offset used in the generated validation data.
}
\examples{

\dontrun{
estimate = AGSI(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
