% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{shape_model_output}
\alias{shape_model_output}
\title{(re)shape model output based upon the class of the input data
and valid model estimates. Mainly, reshapes data to a spatial
raster format when required.}
\usage{
shape_model_output(data, doy)
}
\arguments{
\item{data}{input data generated using the format_*() functions}

\item{doy}{phenophase estimates as a doy value}
}
\value{
raster or vector with estimated phenophase timing (in DOY)
}
\description{
(re)shape model output based upon the class of the input data
and valid model estimates. Mainly, reshapes data to a spatial
raster format when required.
}
\examples{

# Should not be run stand-alone
\dontrun{
shape_model_output(data = data, doy = doy)
}
}
\keyword{model}
\keyword{phenology,}
