% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQ.r
\name{SQ}
\alias{SQ}
\title{Sequential model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)}
\usage{
SQ(par, data)
}
\arguments{
\item{data:}{input data (see reference for detailed description),
data should be formatted using flat_format()}

\item{par:}{a vector of parameter values, this is functions specific}
}
\description{
Sequential model as defined in
Basler et al. 2016 (Agr. For. Meteorlogy)
}
\examples{

\dontrun{
estimate = SQ(data = data, par = par)
}
}
\keyword{model}
\keyword{phenology,}
