% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.pr_fit}
\alias{summary.pr_fit}
\title{Print summary values for fit model output}
\usage{
\method{summary}{pr_fit}(object, ...)
}
\arguments{
\item{object}{input data generated using the pr_fit() function}

\item{...}{additional parameters to pass}
}
\value{
a plot with fit statistics
}
\description{
Print summary values for fit model output
}
\keyword{accuracy}
\keyword{model,}
\keyword{phenology,}
